+++ CHANGES.rst
@@ -1,3 +1,10 @@
+0.9.4
+-----
+- Modified walk() to accept a generic function to filter folders
+- Removed clean_tag(), parse_tags, format_tags()
+- Added utf-7-imap4 codec to parse folder names
+
+
 0.9.3
 -----
 - Fixed revive() to handle messages that lack a date